+++ main.go
@@ -18,7 +18,7 @@ import (
 )
 
 // Set by -ldflags "-X main.version=..."
-var version = "0.3.0"
+var version = "0.4.0"
 
 func main() {
 	if err := run(os.Args[1:]); err != nil {